--- a/src/db/db/db.pro
+++ b/src/db/db/db.pro
@@ -6,6 +6,9 @@ include($$PWD/../../lib.pri)
 
 DEFINES += MAKE_DB_LIBRARY
 
+CONFIG += precompile_header
+PRECOMPILED_HEADER = stable.h
+
 SOURCES = \
   dbArray.cc \
   dbBox.cc \