--- a/src/db/db/stable.h
+++ b/src/db/db/stable.h
@@ -1,0 +1,12 @@
+#include "dbCommon.h"
+
+#include <algorithm>
+#include <set>
+#include <string>
+#include <vector>
+
+#include "gsiClass.h"
+#include "tlObject.h"
+#include "tlString.h"
+
+#include "dbCell.h"