--- a/src/tl/tl/stable.h
+++ b/src/tl/tl/stable.h
@@ -1,0 +1,7 @@
+#include "tlCommon.h"
+
+#include <map>
+#include <string>
+#include <vector>
+
+#include "tlString.h"