@klayoutmatthias 2023-06-22
Thanks for these investigations, but my personal experiences with PCH are not encouraging. KLayout is built on may platforms including very old ones, so I think a 30% improvement in compile time - which is a one time invest after all - does not justify the side effects and increased maintenance effort from my side.

Matthias