@klayoutmatthias 2023-06-22
I did not intend to discourage you ... I hope you did not take it that way :(

To be more precise: while using PCH I found myself wiping the build area more often than I wanted. That happened with MS compilers, but also with gcc versions adopting PCH. Maybe PCH in general are more stable today, but I felt that the header update mechanism do not cover all potential change scenarios and wiping the build area was often the only way to get back to a clean build. By "clean" I mean not showing weird bugs that magically disappear after recompiling. I felt that the build speed advantage does not compensate for these effects. But maybe I should try again with today's compilers.

Matthias