@Partmedia 2023-06-22
No, I understand completely; this is why I did a small proof of concept for consideration before I invested much more effort :)

The header update problem probably has to do with header file ordering that changes the build, e.g. including a config header before a system header may change what gets built. Hence somebody does really need to go through directory by directory and make a PCH for each, instead of having a project-wide PCH.

For development you're probably right that only incremental compile time really matters.